Staging

+++ database/migrations/2019_06_27_056000_create_projects_table.php
@@ -17,8 +17,8 @@ class CreateProjectsTable extends Migration
             $table->date('started_at');
             $table->date('ended_at');
             $table->integer('status');
-            $table->integer('budget');
-            $table->integer('price');
+            $table->bigInteger('budget');
+            $table->bigInteger('price');
             $table->string('file')->nullable();
             $table->timestamps();
